--- a/bench/GEMMsBenchmark.cc
+++ b/bench/GEMMsBenchmark.cc
@@ -115,6 +115,7 @@ void performance_test() {
         NWARMUP,
         NITER,
         flush ? &llc : nullptr);
+    ttot *= 1e9; // convert to ns
 
     ((volatile char*)(llc.data()));